# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateArticle:
    properties:
      description:
        type: string
      topic:
        type: string
    required:
    - description
    - topic
    type: object
...
  dto.CreateComment:
    properties:
      text:
        type: string
    required:
    - text
    type: object
...
  dto.UpdateArticle:
    properties:
      description:
        type: string
      topic:
        type: string
    required:
    - description
    - topic
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend for AITU Corporate Self-Study Portal.
  title: ACSP Backend
  version: "1.0"
paths:
...
  /api/v1/scholar/articles:
    get:
      consumes:
      - application/json
      description: Get all articles in the database
      operationId: get-all-articles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all articles
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Method for creating an article for a user by id in the database
        by user id
      operationId: create-article
      parameters:
      - description: article information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateArticle'
      - description: article image
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create an article
      tags:
      - articles
  /api/v1/scholar/articles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an article from database
      operationId: delete-article-by-id
      parameters:
      - description: article id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete an article by id
      tags:
      - articles
    get:
      consumes:
      - application/json
      description: Get articles by id and user id
      operationId: get-article-by-id-and-user-id
      parameters:
      - description: article id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get article by id and user id
      tags:
      - articles
    put:
      consumes:
      - application/json
      description: Update data of article
      operationId: update-title-by-id
      parameters:
      - description: article id
        in: path
        name: id
        required: true
        type: string
      - description: article information
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateArticle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an article by id
      tags:
      - articles
  /api/v1/scholar/articles/{id}/comments:
    get:
      consumes:
      - application/json
      description: Get all comments of an article
      operationId: get-all-comments-by-article-id
      parameters:
      - description: article id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all comments by article id
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Method for commenting an article by id and user id
      operationId: comment-an-article
      parameters:
      - description: article id
        in: path
        name: id
        required: true
        type: string
      - description: comment information
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateComment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Comment an article
      tags:
      - articles
  /api/v1/scholar/articles/{id}/comments/{commentID}/downvote:
    post:
      consumes:
      - application/json
      description: Downvote comment of an article by article id and comment id
      operationId: downvote-comment
      parameters:
      - description: article id
        in: path
        name: id
        required: true
        type: string
      - description: comment id
        in: path
        name: commentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Downvote comment
      tags:
      - articles
  /api/v1/scholar/articles/{id}/comments/{commentID}/replies:
    get:
      consumes:
      - application/json
      description: Get all comments of an article
      operationId: get-all-replies-of-comment
      parameters:
      - description: article id
        in: path
        name: id
        required: true
        type: string
      - description: comment id
        in: path
        name: commentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all comments by article id
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Reply to comment of an article
      operationId: reply-to-comment-by-article-id-and-comment-id
      parameters:
      - description: article id
        in: path
        name: id
        required: true
        type: string
      - description: comment id
        in: path
        name: commentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Reply to comment by article id and comment id
      tags:
      - articles
  /api/v1/scholar/articles/{id}/comments/{commentID}/upvote:
    post:
      consumes:
      - application/json
      description: Upvote comment of an article by article id and comment id
      operationId: upvote-comment
      parameters:
      - description: article id
        in: path
        name: id
        required: true
        type: string
      - description: comment id
        in: path
        name: commentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Upvote comment
      tags:
      - articles
  /api/v1/scholar/articles/{id}/comments/{commentID}/votes:
    get:
      consumes:
      - application/json
      description: Downvote comment of an article by article id and comment id
      operationId: get-votes-of-comment
      parameters:
      - description: article id
        in: path
        name: id
        required: true
        type: string
      - description: comment id
        in: path
        name: commentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        default:
          description: ""
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get vote difference of comment
      tags:
      - articles
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"